Created Working File path creation for jsondump

+++ jsondump.py
@@ -1,12 +1,14 @@
 import json
+import os
 from pos_converter import position_convert
-from speed_converteer import speed_convert
-from datetime import date
+from speed_converter import speed_convert
+from datetime import datetime
 
 
 class jsondump:
     def __init__(self, lat, lon, sog, cog):
-        self.date = date.today().strftime("%Y-%m-%d")
+        self.date = datetime.today().strftime("%Y-%m-%d")
+        self.time = datetime.now().strftime("%H-%M-%S")
         self.lib = {
             "LAT": float(lat),
             "LON": float(lon),
@@ -15,8 +17,14 @@ class jsondump:
         }
 
     def jsonexport(self):
-        with open("rmc_%s.json" % self.date, "w") as outfile:
-            json.dump(self.lib, outfile)
+        filePath = "../Marine AI Competency Exercise/%s/" % self.date
+        if os.path.exists(filePath):
+            with open("/Marine AI Competency Exercise/%s/rmc_%s.json" % (self.date, self.time), "w") as outfile:
+                json.dump(self.lib, outfile)
+        if not os.path.exists(filePath):
+            os.mkdir(filePath)
+            with open("/Marine AI Competency Exercise/%s/rmc_%s.json" % (self.date, self.time), "w") as outfile:
+                json.dump(self.lib, outfile)
 
 
 if __name__ == "__main__":
